refactor: 테스트 패키지 구조 변경
- 테스트 패키지는 수정한 구조가 반영이 안되어 있어서 변경하였음
