+++ src/test/java/webserver/requesthandler/authenticator/AuthenticatorTest.java
@@ -1,4 +1,4 @@
-package webserver.authenticator;
+package webserver.requesthandler.authenticator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +11,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import webserver.requesthandler.authenticator.Authenticator;
-import webserver.requesthandler.authenticator.UnauthenticatedURLs;
 import webserver.requesthandler.http.HttpRequest;
 import webserver.requesthandler.http.HttpResponse;
 import webserver.requesthandler.http.HttpStatus;
