+++ src/test/java/webserver/requesthandler/handlerimpl/RegistrationTest.java
@@ -1,4 +1,4 @@
-package webserver.requesthandler;
+package webserver.requesthandler.handlerimpl;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -8,7 +8,6 @@ import model.User;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.requesthandler.handlerimpl.RegistrationHandler;
 import webserver.requesthandler.http.HttpRequest;
 import webserver.requesthandler.http.HttpResponse;
