+++ src/test/java/webserver/requesthandler/handlerimpl/RequestHandlerTest.java
@@ -1,4 +1,4 @@
-package webserver.requesthandler;
+package webserver.requesthandler.handlerimpl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,8 +6,6 @@ import java.io.IOException;
 import java.util.Objects;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.requesthandler.handlerimpl.ArticleHandler;
-import webserver.requesthandler.handlerimpl.RequestHandler;
 import webserver.requesthandler.http.HttpResponse;
 
 class RequestHandlerTest {
