+++ src/test/java/webserver/requesthandler/http/ContentTypeTest.java
@@ -1,10 +1,9 @@
-package webserver.http;
+package webserver.requesthandler.http;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.requesthandler.http.ContentType;
 
 class ContentTypeTest {
